Renaming classes with appropriate names as per feedback in class
